LPS-59429 move packages
